chore: half refactor
