+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = ["translatable", "translatable_proc"]
+members = ["translatable", "translatable_proc", "translatable_shared"]
