+++ translatable_proc/Cargo.toml
@@ -15,7 +15,7 @@ proc-macro = true
 proc-macro2 = "1.0.94"
 quote = "1.0.38"
 serde = { version = "1.0.218", features = ["derive"] }
-strum = { version = "0.27.1", features = ["derive"] }
+strum = "0.27.1"
 syn = { version = "2.0.98", features = ["full"] }
 thiserror = "2.0.11"
 toml = "0.8.20"
