+++ translatable_shared/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "translatable_shared"
+description = "Shared dependencies crate for translatable."
+repository = "https://github.com/FlakySL/translatable.rs"
+license = "MIT OR Apache-2.0"
+readme = "../README-MACROS.md"
+version = "0.1.0"
+edition = "2024"
+authors = ["Esteve Autet <esteve@memw.es>", "Chiko <chiko@envs.net>"]
+
+[dependencies]
+strum = { version = "0.27.1", features = ["derive", "strum_macros"] }
